refactor: split `dto` package into `request` and `response` packages to distinguish them later
